Pulled ResultSetMapperDslSupport into its own file.
